requirements.txt
fastapi==0.112.1
uvicorn[standard]==0.30.6
requests==2.32.3
pydantic==2.8.2
numpy==2.0.1
pandas==2.2.2
beautifulsoup4==4.12.3
lxml==5.3.0
matplotlib==3.9.1
google-generativeai==0.7.2
